feat: regist admin and developer
